Add support for u64 types to generator and formatting

+++ cpp/exe/icftool/ConsolePrinter.h
@@ -38,7 +38,7 @@ public:
         std::cout << message << std::endl;
     }
 
-    virtual void print(const char* name, uint32_t value) override
+    virtual void print(const char* name, uint64_t value) override
     {
         this->write_line_prefix();
         std::cout << name << ": " << value << std::endl;
